+++ src/lit.cpp
@@ -60,16 +60,6 @@ bool a2litval(litval_t* val, const char *str)
 	return true;
 }
 
-uint64 bte2mask(bte_t bte)
-{
-	bte_t size = bte & BTE_SIZE_MASK;
-	if(!size || size > 4)
-		size = inf_get_cc_size_e();
-	else
-		size = 1 << (size - 1);
-	return size == 8 ? ~0ULL : ((1ULL << (size * 8)) - 1);
-}
-
 struct ida_local lit_arg_t
 {
 	uint32 n; // number of this argument for function
@@ -227,7 +217,8 @@ public:
 					const char* sp = qstrchr(&buf[2], ' ');
 					litval_t val;
 					if(sp && a2litval(&val, sp + 1)) {
-						refs->insert(std::pair<litval_t, qstring>(val, qstring(&buf[2], sp - &buf[2])));
+						if(!refs->insert(std::pair<litval_t, qstring>(val, qstring(&buf[2], sp - &buf[2]))).second)
+							Log(llFlood, "duplicate literal on line %d\n", line);
 						break;
 					}
 				}
@@ -241,11 +232,13 @@ public:
 
 /*--------------------------------------------------------------------------*/
 static literal_db* lit = NULL;
-static const char* WINERROR = "WINERROR";
+static const char* sWINERROR = "WINERROR";
+static const char* sNTSTATUS = "NTSTATUS";
 
 struct ida_local lit_visitor_t : public ctree_visitor_t
 {
 	cfunc_t *func;
+	tinfo_t return_type;
 	bool cmtModified;
 	user_cmts_t *cmts;
 
@@ -254,6 +247,14 @@ struct ida_local lit_visitor_t : public ctree_visitor_t
 		cmts = restore_user_cmts(cfunc->entry_ea);
 		if(cmts == NULL)
 			cmts = user_cmts_new();
+
+		tinfo_t funcType;
+		if(func->get_func_type(&funcType)) {
+			func_type_data_t fti;
+			if(funcType.get_func_details(&fti) && fti.rettype.is_typeref()) {
+				return_type = fti.rettype;
+			}
+		}
 	}
 
 	~lit_visitor_t()
@@ -263,9 +264,10 @@ struct ida_local lit_visitor_t : public ctree_visitor_t
 		user_cmts_free(cmts);
 	}
 	virtual int idaapi visit_expr(cexpr_t *expr);
+	virtual int idaapi visit_insn(cinsn_t *insn);
 	bool chkCallArg(cexpr_t *expr, qstring &comment);
 	bool chkStrucMemb(cexpr_t *memb, cexpr_t *cons, qstring &comment);
-	bool chkConstType(cexpr_t *expr, cexpr_t *cons, qstring &comment);
+	bool chkConstType(tinfo_t& type, cexpr_t *cons, qstring &comment);
 	cexpr_t* getLiteralExp(cexpr_t *constexp, const literals_t& l, bool exclusive);
 	cexpr_t* makeEnumExpr(const char* name, litval_t val, cexpr_t *constexp);
 };
@@ -299,7 +301,7 @@ static qstring getLiteralString(litval_t val, const literals_t& l, bool exclusiv
 	return str;
 }
 
-static const char* importEnumFromTil(til_t *til, const char* name, litval_t val)
+static const char* importEnumFromTil(til_t *til, const char* name, litval_t val, uchar *serial)
 {
 	enable_numbered_types(til, true);
 
@@ -320,26 +322,57 @@ static const char* importEnumFromTil(til_t *til, const char* name, litval_t val)
 			if (t.deserialize(til, &type, &fields)) {
 				enum_type_data_t ed;
 				if (t.get_enum_details(&ed)) {
-					uint64 mask = bte2mask(ed.bte);
+					uint64 mask = ed.calc_mask();
 					uint64 vm = val & mask;
-					for (auto memb = ed.begin(); memb != ed.end(); memb++) {
-						const char* s1 = nullptr;
-						const char* s2 = nullptr;
-						if (vm == (memb->value & mask) && (!qstrcmp(name, memb->name.c_str()) ||
-							 //HACK GWL_ vs GWLP_ name mismatch. In 64-bit TIL, GWL_USERDATA does not exist - only GWLP_USERDATA.
-							 // value matches but name differs, accept if suffix after 1st '_' matches
-								 (!strncmp(name, "GWL", 3) && (s1 = qstrchr(name, '_'), s2 = qstrchr(memb->name.c_str(), '_'), s1 && s2 && !qstrcmp(s1, s2)))))
-						{
-							const char* typeName = get_numbered_type_name(til, ordinal);
-							if (typeName) {
-#if IDA_SDK_VERSION < 850
-								import_type(til, -1, typeName);
-								Log(llInfo, "import enum \"%s\" from til \"%s\"\n", typeName, til->name);
-#endif //IDA_SDK_VERSION < 850
-								return typeName;
+					*serial = 0;
+					for(auto &memb : ed) {
+						if(vm == (memb.value & mask)) {
+							bool match = !qstrcmp(name, memb.name.c_str());
+							if(!match && !strncmp(name, "GWL", 3)) {
+								//HACK GWL_ vs GWLP_ name mismatch. In 64-bit TIL, GWL_USERDATA does not exist - only GWLP_USERDATA.
+								// value matches but name differs, accept if suffix after 1st '_' matches
+								const char* s1 = qstrchr(name, '_');
+								const char* s2 = qstrchr(memb.name.c_str(), '_');
+								if(!qstrcmp(s1, s2)) {
+									match = true;
+									*serial = 0; //reset serial for hack
+								}
 							}
-							Log(llWarning, "not named enum for \"%s\" 0x%x \n", name, val);
-							return NULL;
+							if(match)	{
+								const char* typeName = get_numbered_type_name(til, ordinal);
+								if (typeName && *typeName) {
+									if(til && til != get_idati()) { // nullptr til is the local til
+#if IDA_SDK_VERSION < 850
+										import_type(til, -1, typeName);
+#else //IDA_SDK_VERSION >= 850
+										//enum doesnt work when type doesnt exist in local til
+#if 0
+										//I've not found a way to import original type, two below
+										//gets correctly named `nt`, but fails on `nt.force_tid();` and emums are not shown
+										tinfo_t nt; nt.get_named_type(til, typeName); nt.force_tid();
+										tinfo_t nt; nt.get_numbered_type(til, ordinal); nt.force_tid();
+#elif 0
+										// bad named but working: desirialized `t` is imported with autogenerated name;
+										if(t.force_tid() != BADADDR) {
+											//this `set_named_type` is not required actually - enums may refer to autogenerated name
+											tinfo_code_t code = t.set_named_type(nullptr, typeName, NTF_NOBASE | NTF_REPLACE);
+											Log(llDebug, "set_named_type('%s') returns %d\n", typeName, code);
+										}
+#else
+										//create a full copy of enum type in local til
+										tinfo_code_t code = t.set_named_type(nullptr, typeName, NTF_TYPE|NTF_REPLACE);
+										if(code != TERR_OK)
+											Log(llDebug, "set_named_type('%s') returns %d\n", typeName, code);
+#endif
+#endif //IDA_SDK_VERSION < 850
+									}
+									Log(llDebug, "found enum '%s' in til '%s', serial:%d val:0x%" FMT_64 "X mask:0x%" FMT_64 "X name:'%s'\n", typeName, til ? til->name : "local", *serial, memb.value, mask,  memb.name.c_str());
+									return typeName;
+								}
+								Log(llWarning, "not named enum ord:%u for \"%s\" 0x%" FMT_64 "X\n", ordinal, name, val);
+								return NULL;
+							}
+							(*serial)++;
 						}
 					}
 				}
@@ -349,16 +382,16 @@ static const char* importEnumFromTil(til_t *til, const char* name, litval_t val)
 	return NULL;
 }
 
-static const char* importEnumFromTils(const char* name, litval_t val)
+static const char* importEnumFromTils(const char* name, litval_t val, uchar *serial)
 {
 	Log(llDebug, "find enum memb %s (0x%x) in tils\n", name, val);
 	til_t *til = (til_t *)get_idati();
-	const char* typeName = importEnumFromTil(til, name, val);
+	const char* typeName = importEnumFromTil(til, name, val, serial);
 	if(typeName)
 		return typeName;
 
 	for (int i = 0; i < til->nbases; i++) {
-			const char* typeName = importEnumFromTil(til->base[i], name, val);
+			const char* typeName = importEnumFromTil(til->base[i], name, val, serial);
 			if (typeName)
 				return typeName;
 	}
@@ -374,7 +407,8 @@ cexpr_t* lit_visitor_t::makeEnumExpr(const char* name, litval_t val, cexpr_t *co
 #if IDA_SDK_VERSION < 850
 	const_t memb = get_enum_member_by_name(name);
 	if(memb == BADNODE) {
-		if(!importEnumFromTils(name, val))
+		uchar ser = 0;
+		if(!importEnumFromTils(name, val, &ser))
 			return NULL;
 		memb = get_enum_member_by_name(name);
 		if(memb == BADNODE)
@@ -390,10 +424,10 @@ cexpr_t* lit_visitor_t::makeEnumExpr(const char* name, litval_t val, cexpr_t *co
 	auto serial = get_enum_member_serial(memb);
 #else //IDA_SDK_VERSION >= 850
 	//FIXME: I've not found fast way to get enum type-name from member-name, maybe need to implement some cashing?
-	const char* typeName = importEnumFromTils(name, val);
+	uchar serial = 0; // it seems this is serial num of member with the same value but different name
+	const char* typeName = importEnumFromTils(name, val, &serial);
 	if(!typeName)
 		return NULL;
-	uchar serial = 0; //FIXME: does it matter???
 	qstring enName = typeName;
 #endif //IDA_SDK_VERSION < 850
 
@@ -406,8 +440,8 @@ cexpr_t* lit_visitor_t::makeEnumExpr(const char* name, litval_t val, cexpr_t *co
 	newexp->n->nf.type_name = enName;
 #if IDA_SDK_VERSION >= 750
 	newexp->n->nf.props |= NF_VALID; // no any other way to set enum, but ida 7.5 raise INTERR 52381 w/o NF_VALID flag
-	newexp->type = create_typedef(enName.c_str()); // ida 7.5 raise INTERR 52378 w/o this
 #endif // IDA_SDK_VERSION >= 750
+	newexp->type = create_typedef(enName.c_str()); // ida 7.5 raise INTERR 52378 w/o this
 	return newexp;
 }
 
@@ -540,14 +574,8 @@ bool lit_visitor_t::chkStrucMemb(cexpr_t *memb, cexpr_t *cons, qstring &comment)
 	return false;
 }
 
-bool lit_visitor_t::chkConstType(cexpr_t *expr, cexpr_t *cons, qstring &comment)
+bool lit_visitor_t::chkConstType(tinfo_t& type, cexpr_t *cons, qstring &comment)
 {
-	tinfo_t type;
-	if (expr->op == cot_cast)
-		type = expr->type;
-	else
-		type = expr->theother(cons)->type;
-
 	qstring typeName;
 	if (!type.get_type_name(&typeName)) {
 		//type.print(&typeName); // for debugging only
@@ -559,7 +587,7 @@ bool lit_visitor_t::chkConstType(cexpr_t *expr, cexpr_t *cons, qstring &comment)
 
 	const literals_t* l = NULL;
 	bool exclusive;
-	if (typeName == WINERROR) {
+	if (typeName == sWINERROR) {
 		lit_func_t flit = lit->find_func("RtlGetLastWin32Error");
 		if (!lit->is_func(flit))
 			return false;
@@ -568,9 +596,16 @@ bool lit_visitor_t::chkConstType(cexpr_t *expr, cexpr_t *cons, qstring &comment)
 			return false;
 		l = &la->refs;
 		exclusive = la->exclusive;
-	}
-#if IDA_SDK_VERSION < 750  //ida 7.5 raise INTERR 52378 (type name conflict? "HANDLE" & "MACRO_INVALID_HANDLE")
-	else if (typeName == "HANDLE") {
+	} else if (typeName == "NTSTATUS") {
+		lit_func_t flit = lit->find_func("ZwClose");
+		if (!lit->is_func(flit))
+			return false;
+		const lit_arg_t *la = lit->find_arg(flit, 0);
+		if (!la)
+			return false;
+		l = &la->refs;
+		exclusive = la->exclusive;
+	} else if (typeName == "HANDLE") {
 		litval_t val = cons->numval();
 		const char* name;
 		switch (val) {
@@ -582,18 +617,12 @@ bool lit_visitor_t::chkConstType(cexpr_t *expr, cexpr_t *cons, qstring &comment)
 			return false;
 		}
 		cexpr_t *newExp = makeEnumExpr(name, val, cons);
-			//IDA bug: prints INVALID_FILE_SIZE instead INVALID_HANDLE_VALUE
 		if (newExp) {
-			newExp->type = expr->type;
-			replaceExp(func, expr, newExp);
+			replaceExp(func, cons, newExp);
 			return true;
 		}
 		return false;
-	}	else {
-		//TODO: add more typenames here
-		return false;
 	}
-#endif // IDA_SDK_VERSION < 750
 	if (!l)
 		return false;
 
@@ -612,21 +641,31 @@ int idaapi lit_visitor_t::visit_expr(cexpr_t *expr)
 	qstring comment;
 	bool changed = false;
 	if(expr->op == cot_call) {
-		changed |= chkCallArg(expr, comment);
-	} else if((expr->op >= cot_asg && expr->op <= cot_ult)) {
+		changed = chkCallArg(expr, comment);
+	} else if((expr->op == cot_asg || is_relational(expr->op))) {
 		cexpr_t *cons = expr->find_op(cot_num);
-		if(cons) {
-			cexpr_t *memb = expr->theother(cons);
-			if(memb->op == cot_memref || memb->op == cot_memptr)
-				changed |= chkStrucMemb(memb, cons, comment);
-			else
-				changed |= chkConstType(expr, cons, comment);
+		if(cons && !cons->n->nf.is_fixed()) {
+			cexpr_t *other = expr->theother(cons);
+			if(other->op == cot_memref || other->op == cot_memptr)
+				changed = chkStrucMemb(other, cons, comment);
+			if(!changed)
+				changed = chkConstType(other->type, cons, comment);
 		}
-	}	else if (expr->op == cot_cast && expr->x->op == cot_num) {
-		changed |= chkConstType(expr, expr->x, comment);
+	}	else if (expr->op == cot_cast && expr->x->op == cot_num && !expr->x->n->nf.is_fixed()) {
+		changed = chkConstType(expr->type, expr->x, comment);
 	}
 
 	cmtModified |= setComment4Exp(func, cmts, expr, comment.c_str(), changed);
+	return 0;
+}
+
+int idaapi lit_visitor_t::visit_insn(cinsn_t *insn)
+{
+	if(insn->op == cit_return && insn->creturn->expr.op == cot_num && !insn->creturn->expr.n->nf.is_fixed() && !return_type.empty()) {
+		qstring comment;
+		bool changed = chkConstType(return_type, &insn->creturn->expr, comment);
+		cmtModified |= setComment4Exp(func, cmts, &insn->creturn->expr, comment.c_str(), changed);
+	}
 	return 0;
 }
 
@@ -647,7 +686,7 @@ bool lit_overrideTypes()
 	if (!lit)
 		return true;
 
-	if (get_named_type(NULL, WINERROR, NTF_TYPE | NTF_NOBASE))
+	if (get_named_type(NULL, sWINERROR, NTF_TYPE | NTF_NOBASE))
 		return true;
 
 	uint32 ord = alloc_type_ordinal(NULL);
@@ -656,12 +695,12 @@ bool lit_overrideTypes()
 
 	tinfo_t t;
 	t.create_simple_type(BT_INT32);
-	t.set_named_type(NULL, WINERROR, 0);
+	t.set_named_type(NULL, sWINERROR, 0);
 
 	const char* const funcnames[] = {"GetLastError", "__imp_GetLastError", "RtlGetLastWin32Error", "RegOpenKey"};
 	for(size_t i = 0, n = qnumber(funcnames); i < n; i++) {
 		ea_t ea = get_name_ea(BADADDR, funcnames[i]);
-		if (ea == BADADDR /*|| is_func(get_flags(ea))*/)
+		if (ea == BADADDR)
 			continue;
 		tinfo_t ft;
 		get_tinfo(&ft, ea);